move arith routines to arith.cpp
