+++ ext/mri/ruby_arrayfire.cpp
@@ -44,19 +44,7 @@ namespace arf {
     }
   }
 
-  static void hostArray(afstruct *afarray)
-  {
-
-  }
-
-  static void add(afstruct *result, afstruct *left, afstruct *right)
-  {
-    array l = array(left->dimension[0], left->dimension[1], left->array);
-    array r = array(right->dimension[0], right->dimension[1], right->array);
-    array res = operator+(l,r);
-    result->array = res.host<double>();
-  }
-
+  #include "arith.cpp"
   #include "blas.cpp"
   #include "lapack.cpp"
 }
